allow message formatting in events, renumbered applemidi events
git-svn-id: https://midikit.googlecode.com/svn/trunk@95 c9dcb2c1-a6a4-760b-d948-6d6a4c2e33b0

+++ midi/event.h
@@ -5,7 +5,7 @@
 struct MIDIEvent;
 extern struct MIDITypeSpec * MIDIEventType;
 
-struct MIDIEvent * MIDIEventCreate( int id, void * info, char * message, ... );
+struct MIDIEvent * MIDIEventCreate( size_t id, void * info, char * message, ... );
 void MIDIEventDestroy( struct MIDIEvent * event );
 void MIDIEventRetain( struct MIDIEvent * event );
 void MIDIEventRelease( struct MIDIEvent * event );
